+++ apps/assets/api/__init__.py
@@ -1,9 +1,7 @@
 from .mixin import *
-from .admin_user import *
 from .asset import *
 from .label import *
 from .system_user import *
-from .system_user_relation import *
 from .accounts import *
 from .node import *
 from .domain import *
