+++ apps/assets/models/__init__.py
@@ -2,7 +2,6 @@ from .base import *
 from .asset import *
 from .label import Label
 from .user import *
-from .cluster import *
 from .group import *
 from .domain import *
 from .node import *
@@ -12,5 +11,5 @@ from .utils import *
 from .authbook import *
 from .gathered_user import *
 from .favorite_asset import *
-from .backup import *
 from .account import *
+from .backup import *
