+++ apps/assets/serializers/__init__.py
@@ -2,7 +2,6 @@
 #
 
 from .asset import *
-from .admin_user import *
 from .label import *
 from .system_user import *
 from .node import *
