+++ apps/assets/signal_handlers/__init__.py
@@ -1,5 +1,4 @@
 from .asset import *
-from .system_user import *
-from .authbook import *
+from .account import *
 from .node_assets_amount import *
 from .node_assets_mapping import *
