+++ apps/perms/models/asset_permission.py
@@ -22,14 +22,13 @@ logger = logging.getLogger(__name__)
 class AssetPermission(BasePermission):
     assets = models.ManyToManyField('assets.Asset', related_name='granted_by_permissions', blank=True, verbose_name=_("Asset"))
     nodes = models.ManyToManyField('assets.Node', related_name='granted_by_permissions', blank=True, verbose_name=_("Nodes"))
-    system_users = models.ManyToManyField('assets.SystemUser', related_name='granted_by_permissions', blank=True, verbose_name=_("System user"))
+    accounts = models.JSONField(default=list, verbose_name=_("Accounts"))
 
     class Meta:
         unique_together = [('org_id', 'name')]
         verbose_name = _("Asset permission")
         ordering = ('name',)
-        permissions = [
-        ]
+        permissions = []
 
     @lazyproperty
     def users_amount(self):
@@ -46,10 +45,6 @@ class AssetPermission(BasePermission):
     @lazyproperty
     def nodes_amount(self):
         return self.nodes.count()
-
-    @lazyproperty
-    def system_users_amount(self):
-        return self.system_users.count()
 
     @classmethod
     def get_queryset_with_prefetch(cls):
@@ -78,10 +73,6 @@ class AssetPermission(BasePermission):
 
     def assets_display(self):
         names = [asset.hostname for asset in self.assets.all()]
-        return names
-
-    def system_users_display(self):
-        names = [system_user.name for system_user in self.system_users.all()]
         return names
 
     def nodes_display(self):
